pyproject.toml
[project]
name = "Furina"
version = "1.0.0"
description = "A discord.py bot i made while learning python"
authors = [{name = "Nguyễn Duy Thanh", email = "thvnhz@gmail.com"}]
license = "Apache-2.0"
dependencies = [
    "aiohttp",
    "asqlite",
    "dateparser",
    "discord-py",
    "docstring-parser",
    "enka",
    "jishaku",
    "numpy",
    "psutil",
    "pynacl",
    "python-dotenv",
    "tqdm",
    "lavalink",
    "requests",
    "typing-extensions>=4.14.1",
    "anyio>=4.11.0",
]

requires-python = "<4.0.0,>=3.10.0"

[tool.setuptools]
packages = ['core', 'cogs', 'assets']


[tool.ruff]
line-length = 80
target-version = "py310"

[tool.ruff.lint]
# Config by Umbra @ https://github.com/AbstractUmbra/Mipha/blob/main/pyproject.toml
preview = true
select = [
    "A",
    "ANN",
    "ASYNC",
    "B",
    "BLE",
    "C4",
    "COM",
    "DTZ",
    "E",
    "EM",
    "ERA",
    "F",
    "FA",
    "FBT",
    "FURB",
    "G",
    "I",
    "INP",
    "ISC",
    "NPY",
    "PD",
    "PERF",
    "PGH",
    "PIE",
    "PLC",
    "PLE",
    "PLW",
    "PTH",
    "PYI",
    "Q",
    "Q003",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "SLOT",
    "T20",
    "TC",
    "TID",
    "TRY",
    "UP",
    "YTT"
]
ignore = [
    "ANN401",
    "ASYNC109", # dpy commonly uses this parameter
    "COM812",
    "EM101",    # in this case this is okay
    "F401",
    "F402",
    "F403",
    "F405",
    "INP001",   # due to how I load modules this results in annoyances
    "PD011",    # this is not a numpy codebase
    "PERF203",
    "PLC0414",  # pyright ruling for `as` imports needed
    "Q000",
    "RUF001",
    "RUF009",
    "S101",     # assert is used for type checking in this codebase
    "SIM105",
    "TRY003",   # over-eager rule
    "TRY301",   # unrealistic rule
    "UP034",
]
unfixable = [
    "E501", # line length handled in other ways by ruff format
    "ERA",  # Don't delete commented out code
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.isort]
split-on-trailing-comma = true
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false
parametrize-names-type = "csv"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[tool.ruff.lint.flake8-tidy-imports.banned-api]
# https://discuss.python.org/t/problems-with-typeis/55410/6
# https://discuss.python.org/t/problems-with-typeis/55410/46
# Until what can go into a TypeIs/TypeGuard changes, these are just dangerous.
"typing.TypeIs".msg = "TypeIs is fundamentally unsafe, even when using it as described to be safe"
"typing.TypeGuard".msg = "TypeGuard is fundamentally unsafe"
"typing_extensions.TypeIs".msg = "TypeIs is fundamentally unsafe, even when using it as described to be safe"
"typing_extensions.TypeGuard".msg = "TypeGuard is fundamentally unsafe"

